add feature to change password and UI fix

+++ app/src/main/java/com/cs175/bulletinandroid/bulletin/Tabs/Tab5.java
@@ -14,22 +14,31 @@ import android.os.ParcelFileDescriptor;
 import android.provider.MediaStore;
 import android.support.v4.app.Fragment;
 import android.support.v7.app.AlertDialog;
+import android.text.method.PasswordTransformationMethod;
+import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.WindowManager;
 import android.widget.Button;
+import android.widget.EditText;
 import android.widget.ImageView;
+import android.widget.RelativeLayout;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import com.cs175.bulletinandroid.bulletin.BulletinSingleton;
 import com.cs175.bulletinandroid.bulletin.Elements.AlertDialogController;
 import com.cs175.bulletinandroid.bulletin.Elements.RoundedImageView;
+import com.cs175.bulletinandroid.bulletin.FormatValidator;
 import com.cs175.bulletinandroid.bulletin.LoginActivityBackground;
 import com.cs175.bulletinandroid.bulletin.OnRequestListener;
 import com.cs175.bulletinandroid.bulletin.R;
+import com.cs175.bulletinandroid.bulletin.RegisterActivity;
 import com.cs175.bulletinandroid.bulletin.RegistrationCompletedActivity;
 import com.cs175.bulletinandroid.bulletin.Response;
+
+import org.w3c.dom.Text;
 
 import java.io.FileDescriptor;
 import java.io.IOException;
@@ -52,6 +61,13 @@ public class Tab5 extends Fragment implements OnRequestListener {
     private static final int RESULT_LOAD_IMAGE = 1;
     private static final int CAMERA_REQUEST = 1888;
     private AlertDialogController alertDialog;
+    private TextView emailText;
+    private EditText passwordText;
+    private EditText confirmText;
+    private EditText hiddenText;
+    private FormatValidator validator;
+    private TextView mainTitle;
+
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
 
@@ -61,6 +77,90 @@ public class Tab5 extends Fragment implements OnRequestListener {
         settingPicture = (RoundedImageView)view.findViewById(R.id.SettingPicture);
         alertDialog = new AlertDialogController();
 
+        emailText = (TextView)view.findViewById(R.id.SettingEmailText);
+        passwordText = (EditText)view.findViewById(R.id.SettingPasswordText);
+        confirmText = (EditText)view.findViewById(R.id.SettingConfirmText);
+        hiddenText = (EditText)view.findViewById(R.id.SettingHiddenText);
+        mainTitle = (TextView)view.findViewById(R.id.SettingMainTitle);
+
+        mainTitle.setTypeface(singleton.getInstance().getFont());
+
+        hiddenText.requestFocus();
+        passwordText.setTransformationMethod(new PasswordTransformationMethod());
+        confirmText.setTransformationMethod(new PasswordTransformationMethod());
+        validator = new FormatValidator();
+
+        passwordText.setOnEditorActionListener(new TextView.OnEditorActionListener() {
+            @Override
+            public boolean onEditorAction(TextView v, int actionId, KeyEvent event) {
+
+                int index = validator.validatePassword(passwordText.getText().toString());
+
+                if (index == 0) {
+                    confirmText.requestFocus();
+                    return true;
+                } else if (index == 1) {
+                    alertDialog.showDialog(getActivity(), "Password is too short!");
+                    return false;
+                } else if (index == 2) {
+                    alertDialog.showDialog(getActivity(), "Password is too long!");
+                    return false;
+                }
+
+                return false;
+            }
+        });
+
+        hiddenText.setOnFocusChangeListener(new View.OnFocusChangeListener() {
+            @Override
+            public void onFocusChange(View v, boolean hasFocus) {
+                if(hasFocus) {
+                    settingButton.setText("Logout");
+                }
+            }
+        });
+
+
+
+
+        passwordText.setOnFocusChangeListener(new View.OnFocusChangeListener() {
+            @Override
+            public void onFocusChange(View v, boolean hasFocus) {
+                if(hasFocus){
+                    settingButton.setText("Next");
+                }
+            }
+        });
+
+        confirmText.setOnEditorActionListener(new TextView.OnEditorActionListener() {
+            @Override
+            public boolean onEditorAction(TextView v, int actionId, KeyEvent event) {
+                String password = confirmText.getText().toString();
+                if (password.equals(passwordText.getText().toString())) {
+
+                    singleton.getInstance().getAPI().updatePassword((OnRequestListener) getActivity(), password);
+                    hiddenText.requestFocus();
+                    return true;
+                } else {
+                    alertDialog.showDialog(getActivity(), "Passwords do not match!");
+                }
+                return false;
+            }
+        });
+
+        confirmText.setOnFocusChangeListener(new View.OnFocusChangeListener() {
+            @Override
+            public void onFocusChange(View v, boolean hasFocus) {
+                if(hasFocus){
+                    settingButton.setText("Confirm");
+                }
+            }
+        });
+
+
+
+
+        emailText.setText(singleton.getInstance().getUserResponse().getEmail());
         getActivity().getWindow().setSoftInputMode(WindowManager.LayoutParams.SOFT_INPUT_ADJUST_PAN);
 
 
@@ -71,13 +171,15 @@ public class Tab5 extends Fragment implements OnRequestListener {
         //alertDialog.showDialog(getActivity(), singleton.getInstance().getUserResponse().getProfile_picture());
         viewHolder.url = singleton.getInstance().getUserResponse().getProfile_picture();
         settingPicture.setTag(viewHolder);
-        new DownloadAsyncTask().execute(viewHolder);
+        new DownloadAsyncTaskProfilePicture().execute(viewHolder);
 
 
 
         settingPicture.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
+
+
                 CharSequence colors[] = new CharSequence[] {"Take a picture", "Choose from gallery"};
 
                 AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
@@ -104,14 +206,44 @@ public class Tab5 extends Fragment implements OnRequestListener {
         settingButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                SharedPreferences.Editor editor = getActivity().getSharedPreferences(GET_TOKEN, MODE_PRIVATE).edit();
-                editor.clear();
-                editor.commit();
-                Intent intent = new Intent(getActivity(), LoginActivityBackground.class);
-                startActivity(intent);
-                getActivity().finish();
+                String title = settingButton.getText().toString();
+                if (title.equals("Logout")) {
+                    SharedPreferences.Editor editor = getActivity().getSharedPreferences(GET_TOKEN, MODE_PRIVATE).edit();
+                    editor.clear();
+                    editor.commit();
+                    Intent intent = new Intent(getActivity(), LoginActivityBackground.class);
+                    startActivity(intent);
+                    getActivity().finish();
+
+                } else if (title.equals("Next")) {
+
+                    int index = validator.validatePassword(passwordText.getText().toString());
+
+                    if (index == 0) {
+
+                        confirmText.requestFocus();
+                    } else if (index == 1) {
+                        alertDialog.showDialog(getActivity(), "Password is too short!");
+
+                    } else if (index == 2) {
+                        alertDialog.showDialog(getActivity(), "Password is too long!");
+
+                    }
+
+                } else if (title.equals("Confirm")) {
+
+                    String password = confirmText.getText().toString();
+                    if (password.equals(passwordText.getText().toString())) {
+
+                        singleton.getInstance().getAPI().updatePassword((OnRequestListener) getActivity(), password);
+                        hiddenText.requestFocus();
+                    } else {
+                        alertDialog.showDialog(getActivity(), "Passwords do not match!");
+                    }
+                }
             }
         });
+
 
         return view;
 
@@ -124,6 +256,7 @@ public class Tab5 extends Fragment implements OnRequestListener {
         if (requestCode == CAMERA_REQUEST && resultCode == RESULT_OK) {
             Bitmap photo = (Bitmap) data.getExtras().get("data");
             settingPicture.setImageBitmap(photo);
+            singleton.getInstance().setflag("tab5");
             singleton.getInstance().getAPI().uploadImage((OnRequestListener)getActivity(), photo);
 
         }
@@ -149,9 +282,8 @@ public class Tab5 extends Fragment implements OnRequestListener {
                 e.printStackTrace();
             }
             settingPicture.setImageBitmap(bmp);
-
+            singleton.getInstance().setflag("tab5");
             singleton.getInstance().getAPI().uploadImage((OnRequestListener)getActivity(), bmp);
-
 
 
         }
